Fixed #43 -- Use prefetch_related when querying tags
This commit changes the internal behaviour of how cosinnus fetches the
tags for taggable objects from the database and passes them to the
render context. Do make the querying more efficient an update of
django-taggit >= 0.11 is required!

Update your environments accordingly!

This commit will also increase the minor version: v0.2