Add tests for swappable user profile
Django cannot handle tests where both a swappable and a swapped model
are used if they define a ForeignKey pointing to the same other model
and have the same related name. Thus the suggestion by a Django core
developer is to add a custom settings module an run the tests with both,
the default model and the swapped model.