Add core support for django-multiform to be used for the tag object
When working on a more sophisticated TagObject on another project, some
shortcomings have been identified in the TagObjectFormMixin. Although
this mixin works fine with trivial TagObjectModels, there is now way to
support special handling of some fields on the form. django-multiform
poses a wrapper to multiple inner forms and provides the same API to the
outside as a ModelForm has. The idea is to make a Multi(Model)Form
directly usabe in Django's generic class based views.