Fix guest account creation DB race condition
Not really a "race condition" as such, but for certain signals (membership update trigger hooks) to work, the user instance must be persisted to the DB before group memberships can be safely added.
In this case, the BBB MembershipUpdateHookThread could not assign a freshly created guest user to a BBB room because it hadn't been saved yet.

https://git.wechange.de/gl/code/portals/fzs/-/issues/22