v1.20.9: Do not consider unverified email user accounts "not active"
This prevented their profile being visible and being able to be found in autocompletes. Also showed them as "deactivated" in user lists.

Undo change from d2efef6a6f55b89f21eb8b9093a8deb05f431497
This was a mistaken change, as the check for __unverified__ emails is a legacy check and was not meant to check for unverified email users.