Deck app and integration (#529)
* Add cosinnus_deck app

- register new app
- add deck view

* Add deck id group field

* Add deck id to group admin

- add deck id to group admins
- exclude nextcloud and deck fields if disabled
- reuse excluded project fields in society admin

* Add nextcloud group initialized signal

Used to create the group deck board

* Add deck connection

* Add helper to check if an integrated app is enabled in a group

* Add integration hooks for group app activated and deactivated

* Add deck integration handler

* Only save relevant fields in nextcloud hooks

Saving the whole group leads to overwriting the group deck id that is set in parallel.

* Improve deck view

- add deck id to div
- try creating the board if not yet created

* Add COSINNUS_DECK_ENABLED setting

* Add support for groups with deck app but without cloud app

Split the nextcloud hooks such that the nextcloud group sync is done if the deck or the cloud app is enabled.
The nextcloud folder sync is done additionally if the cloud app is enabled.

* Remove share permissions for decks

* Fix migration

* Add group permissions to main contnet

Adding space element to main content data containing group permissions.

* Add translation for deck app