#447 Improves BBB recordings access logic (#537)
* #447 Sets group flag for potential BBB recordings

Adds a flag to the source-group when BBB recordings are enabled during room creation.

https://git.wechange.de/gl/code/cooperation/-/issues/447

* #447 Extends BBB recordings access check

- extend group access check for BBB recordings to check for the `may_have-recordings` flag in `group.settings`
- add more explanations to docstring

https://git.wechange.de/gl/code/cooperation/-/issues/447

* #447 Simplifies BBB recordings access check

Removes a redundant check for BBB availability based on `premium_features_expired_on` in `group.settings` because it is superseded by determined by `may_have_bbb_recordings` which includes recordings from event-meetings

Note: the old flag is still set by the cron-job, since it serves as context-info for admins

https://git.wechange.de/gl/code/cooperation/-/issues/447

* Revert "#447 Simplifies BBB recordings access check"

This reverts commit 6058c01a337612a63a25cadd9d04a97ba1a1b1eb.

Removing this check would potentially hide the recorded meetings entry for premium groups that already have recorded meetings and where no meeting was started in the meantime.

---------

Co-authored-by: Sascha Sommer <saschanarr@gmail.com>